requirements.txt
Django==4.2.23
djangorestframework>=3.12.4
pillow==11.3.0
